Update README.md
Documentation improvements according current state of matter e.g. Console is the debugger System console and not the the device one nanoFramework.Console. Additional notes added for the newbies to have an easier start. Changes on all screen initialization and  Console related code tested on a StickCPlus....screen initialization for Core2 and StickC must be added yet.